Create CLI package for user to start services.
This commit creates a CLI package which segregates the different types
of 'services' used to compose the audmon functionality. Those currently
consist of a daemon, server, and client which CLI commands respectively
named as such. Used spf13/cobra package to implement the CLI and flags.
No real functionality is implemented in this commit and the various CLI
commands still only do the dummy Printer() function.

* add dep on spf13/cobra for cli impl
* add pkg/cli to project to introduce service start commands