Add and establish basic websocket communication.
This commit stubs out and exercises the basic websocket functionality
using example code from the gorilla/websocket pkg to demonstrate. The
patterns here are meant to help flesh out the longer term design of
the program and are not meant to be a final approach to the longer
running client and server processes.

* Adds basic server websocket functionality
* Adds basic client websocket functionality
* Adds examples for basic UI from gorilla/websocket example