Add config package to project.
This commit adds a basic configuraiton package to the project under
pkg/config. This package is responsible for reading default config
from memory and writing them to a file, should no config file exist
already. If a configuraiton file does exist, that will be preferred
over the use of in memory defaults.

* Add config package
* Add file checks and write for config file
* Add default config for logging mechanism