Update README, Makefile, and go mod files.
This commit updates the README.md of the root project to include
information about building and running the audmon binary as well
as updating project goals for config and logging. Makefile updates
include adding some boiler plate and build target for ease of use.
Go mod files are updated to include third party dependencies for
sirupsen/logrus and moorgar/venom.

* Update README.md
* Update Makefile to include build target
* Go mod files updated for sirupsen/logrus and moorgar/venom