Basic usage of writing audit events from kernel to log file.
This commit is a trivial exercise in writing audit events from the
auditd kernel messages to a configured log file. This will not be
the ultimate functionality (unless user desires) as the events will
be pushed over a unix socket to the server where they will be sent
to the client on the other end of the websocket connecitons.

* Adds trivial example of writing audit daemon messages to log file