Set up structure for independent daemon binary.
The eventual structure of this project will be similar to that of
programs like Docker or Ogre where there is a cli binary as well as
a daemon binary. The cli will be able to start and stop the daemon
as well as send commands. This commit introduces the beginnings of
this project structure.

* Stub daemon command for second binary.